requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
motor
python-dotenv
httpx